add a destroy_account endpoint
destroy account
- moves over files that the user whishes to move over
- deletes the user

+++ apps/core-api/lib/api/accounts/file.ex
@@ -55,19 +55,17 @@ defmodule Api.Accounts.File do
   def delete_attachment(file) do
     cdn_base_url = System.get_env("UGC_S3_COMPAT_CDN_BASE_URL") || " "
 
-    case String.starts_with?(file.remote_location, cdn_base_url) do
-      true ->
+    if String.starts_with?(file.remote_location, cdn_base_url) do
+      Task.start(fn ->
         file.remote_location
         |> String.replace_leading(cdn_base_url, "")
         |> String.replace_leading("/", "")
         |> String.split("/", parts: 2)
         |> List.last()
         |> UploadService.delete_from_space()
-
-        {:ok, file}
-
-      _ ->
-        {:ok, file}
+      end)
     end
+
+    {:ok, file}
   end
 end
